use usize instead of i32, and remove unnecessary comments

+++ src/main.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, Copy)]
 struct Item {
-    weight: i32,
-    value: i32,
+    weight: usize,
+    value: usize,
 }
 
 struct Knapsack {
-    max_value: i32,
+    max_value: usize,
     selected_items: Vec<Item>,
 }
 
@@ -26,52 +26,46 @@ fn main() {
     println!("Selected items: {:?}", selected_items);
 }
 
-fn knapsack(items: &[Item], capacity: i32) -> Knapsack {
+fn knapsack(items: &[Item], capacity: usize) -> Knapsack {
     let n = items.len();
-    let mut dp = vec![vec![0; (capacity + 1) as usize]; n + 1];
+    let mut dp = vec![vec![0; capacity + 1]; n + 1];
 
-    // Print initial DP table (all zeros)
     println!("Initial DP table:");
     print_dp_table(&dp, capacity);
 
-    // Fill the DP table
     for i in 1..=n {
         for j in 0..=capacity {
             if items[i - 1].weight <= j {
-                dp[i][j as usize] = std::cmp::max(
-                    dp[i - 1][j as usize],
-                    dp[i - 1][(j - items[i - 1].weight) as usize] + items[i - 1].value,
+                dp[i][j] = std::cmp::max(
+                    dp[i - 1][j],
+                    dp[i - 1][j - items[i - 1].weight] + items[i - 1].value,
                 );
             } else {
-                dp[i][j as usize] = dp[i - 1][j as usize];
+                dp[i][j] = dp[i - 1][j];
             }
         }
-        // Print DP table after processing each item
         println!("\nDP table after processing item {} (weight: {}, value: {}):", i, items[i - 1].weight, items[i - 1].value);
         print_dp_table(&dp, capacity);
     }
 
-    // Backtrack to find the selected items
     let mut selected_items = Vec::new();
     let mut remaining_capacity = capacity;
     for i in (1..=n).rev() {
-        if dp[i][remaining_capacity as usize] != dp[i - 1][remaining_capacity as usize] {
+        if dp[i][remaining_capacity] != dp[i - 1][remaining_capacity] {
             selected_items.push(items[i - 1]);
             remaining_capacity -= items[i - 1].weight;
         }
     }
 
-    // Reverse to maintain the original order of items
     selected_items.reverse();
 
     Knapsack {
-        max_value: dp[n][capacity as usize],
+        max_value: dp[n][capacity],
         selected_items,
     }
 }
 
-// Helper function to print the DP table
-fn print_dp_table(dp: &[Vec<i32>], capacity: i32) {
+fn print_dp_table(dp: &[Vec<usize>], capacity: usize) {
     print!("     ");
     for j in 0..=capacity {
         print!("{:2} ", j);
